change files' names (calendars) = consistency
