+++ app/components/UI/calendar/Calendar.js
@@ -1,103 +1,120 @@
-export default function Calendar(props) {
-  const ARRAY = [];
+import * as React from "react";
+import dayjs from "dayjs";
+import Badge from "@mui/material/Badge";
+import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
+import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
+import { PickersDay } from "@mui/x-date-pickers/PickersDay";
+import { DateCalendar } from "@mui/x-date-pickers/DateCalendar";
+import { DayCalendarSkeleton } from "@mui/x-date-pickers/DayCalendarSkeleton";
 
-  for (let i = 0; i < props.array.length; i++) {
-    ARRAY.push(props.array[i]);
-  }
+function getRandomNumber(min, max) {
+  return Math.round(Math.random() * (max - min) + min);
+}
+
+/**
+ * Mimic fetch with abort controller https://developer.mozilla.org/en-US/docs/Web/API/AbortController/abort
+ * ⚠️ No IE11 support
+ */
+function fakeFetch(date, { signal }) {
+  return new Promise((resolve, reject) => {
+    const timeout = setTimeout(() => {
+      const daysInMonth = date.daysInMonth();
+      const daysToHighlight = [1, 2, 3].map(() =>
+        getRandomNumber(1, daysInMonth)
+      );
+
+      resolve({ daysToHighlight });
+    }, 500);
+
+    signal.onabort = () => {
+      clearTimeout(timeout);
+      reject(new DOMException("aborted", "AbortError"));
+    };
+  });
+}
+
+const initialValue = dayjs("2022-04-17");
+
+function ServerDay(props) {
+  const { highlightedDays = [], day, outsideCurrentMonth, ...other } = props;
+
+  const isSelected =
+    !props.outsideCurrentMonth &&
+    highlightedDays.indexOf(props.day.date()) >= 0;
 
   return (
-    <div>
-      {ARRAY.map((event) => {
-        return (
-          <div>
-            <div class="p-5 mb-4 border border-gray-100 rounded-lg bg-gray-50 dark:bg-gray-800 dark:border-gray-700">
-              <time class="text-lg font-semibold text-gray-900 dark:text-white">
-                {event.month}
-              </time>
-              <ol class="mt-3 divide-y divider-gray-200 dark:divide-gray-700">
-                <li>
-                  <a
-                    href={event.href}
-                    class="items-center block p-3 sm:flex hover:bg-gray-100 dark:hover:bg-gray-700"
-                  >
-                    <img
-                      class="w-12 h-12 mb-3 mr-3 rounded-full sm:mb-0"
-                      src={event.imgURL}
-                      alt={event.title}
-                    />
-                    <div class="text-gray-600 dark:text-gray-400">
-                      <div class="text-base font-normal">
-                        <span class="font-medium text-gray-900 dark:text-white"></span>{" "}
-                        {event.title} <p> </p>
-                        <span class="font-medium text-gray-900 dark:text-white">
-                          {event.district} || {event.town2}
-                        </span>{" "}
-                      </div>
-                      <div class="text-sm font-normal">{event.description}</div>
-                      <span class="inline-flex items-center text-xs font-normal text-gray-500 dark:text-gray-400">
-                        <svg
-                          class="w-2.5 h-2.5 mr-1"
-                          aria-hidden="true"
-                          xmlns="http://www.w3.org/2000/svg"
-                          fill="currentColor"
-                          viewBox="0 0 20 20"
-                        >
-                          <path d="M10 .5a9.5 9.5 0 1 0 0 19 9.5 9.5 0 0 0 0-19ZM8.374 17.4a7.6 7.6 0 0 1-5.9-7.4c0-.83.137-1.655.406-2.441l.239.019a3.887 3.887 0 0 1 2.082 2.5 4.1 4.1 0 0 0 2.441 2.8c1.148.522 1.389 2.007.732 4.522Zm3.6-8.829a.997.997 0 0 0-.027-.225 5.456 5.456 0 0 0-2.811-3.662c-.832-.527-1.347-.854-1.486-1.89a7.584 7.584 0 0 1 8.364 2.47c-1.387.208-2.14 2.237-2.14 3.307a1.187 1.187 0 0 1-1.9 0Zm1.626 8.053-.671-2.013a1.9 1.9 0 0 1 1.771-1.757l2.032.619a7.553 7.553 0 0 1-3.132 3.151Z" />
-                        </svg>
-                        {event.region}
-                      </span>
-                    </div>
-                  </a>
-                </li>
-              </ol>
-            </div>
-            <div class="p-5 border border-gray-100 rounded-lg bg-gray-50 dark:bg-gray-800 dark:border-gray-700">
-              <time class="text-lg font-semibold text-gray-900 dark:text-white">
-                Mais detalhes
-              </time>
-              <ol class="mt-3 divide-y divider-gray-200 dark:divide-gray-700">
-                <li>
-                  <a
-                    href="#"
-                    class="items-center block p-3 sm:flex hover:bg-gray-100 dark:hover:bg-gray-700"
-                  >
-                    <img
-                      class="w-12 h-12 mb-3 mr-3 rounded-full sm:mb-0"
-                      src={event.imgURL}
-                      alt="Laura Romeros image"
-                    />
-                    <div class="text-gray-600 dark:text-gray-400">
-                      <div class="text-base font-normal">
-                        <span class="font-medium text-gray-900 dark:text-white">
-                          {event.paragraph1} <p> </p>
-                        </span>{" "}
-                      </div>
-                      <div class="text-sm font-normal">
-                        {event.paragraph2} <p> </p>
-                        {event.paragraph3}
-                      </div>
-                      <span class="inline-flex items-center text-xs font-normal text-gray-500 dark:text-gray-400">
-                        <svg
-                          class="w-2.5 h-2.5 mr-1"
-                          aria-hidden="true"
-                          xmlns="http://www.w3.org/2000/svg"
-                          fill="currentColor"
-                          viewBox="0 0 20 20"
-                        >
-                          <path d="m2 13.587 3.055-3.055A4.913 4.913 0 0 1 5 10a5.006 5.006 0 0 1 5-5c.178.008.356.026.532.054l1.744-1.744A8.973 8.973 0 0 0 10 3C4.612 3 0 8.336 0 10a6.49 6.49 0 0 0 2 3.587Z" />
-                          <path d="m12.7 8.714 6.007-6.007a1 1 0 1 0-1.414-1.414L11.286 7.3a2.98 2.98 0 0 0-.588-.21l-.035-.01a2.981 2.981 0 0 0-3.584 3.583c0 .012.008.022.01.033.05.204.12.401.211.59l-6.007 6.007a1 1 0 1 0 1.414 1.414L8.714 12.7c.189.091.386.162.59.211.011 0 .021.007.033.01a2.981 2.981 0 0 0 3.584-3.584c0-.012-.008-.023-.011-.035a3.05 3.05 0 0 0-.21-.588Z" />
-                          <path d="M17.821 6.593 14.964 9.45a4.952 4.952 0 0 1-5.514 5.514L7.665 16.75c.767.165 1.55.25 2.335.251 6.453 0 10-5.258 10-7 0-1.166-1.637-2.874-2.179-3.407Z" />
-                        </svg>
-                        {event.region}
-                      </span>
-                    </div>
-                  </a>
-                </li>
-              </ol>
-            </div>
-          </div>
-        );
-      })}
-    </div>
+    <Badge
+      key={props.day.toString()}
+      overlap="circular"
+      badgeContent={isSelected ? "🌚" : undefined}
+    >
+      <PickersDay
+        {...other}
+        outsideCurrentMonth={outsideCurrentMonth}
+        day={day}
+      />
+    </Badge>
+  );
+}
+
+export default function Calendar() {
+  const requestAbortController = React.useRef(null);
+  const [isLoading, setIsLoading] = React.useState(false);
+  const [highlightedDays, setHighlightedDays] = React.useState([1, 2, 15]);
+
+  const fetchHighlightedDays = (date) => {
+    const controller = new AbortController();
+    fakeFetch(date, {
+      signal: controller.signal,
+    })
+      .then(({ daysToHighlight }) => {
+        setHighlightedDays(daysToHighlight);
+        setIsLoading(false);
+      })
+      .catch((error) => {
+        // ignore the error if it's caused by `controller.abort`
+        if (error.name !== "AbortError") {
+          throw error;
+        }
+      });
+
+    requestAbortController.current = controller;
+  };
+
+  React.useEffect(() => {
+    fetchHighlightedDays(initialValue);
+    // abort request on unmount
+    return () => requestAbortController.current?.abort();
+  }, []);
+
+  const handleMonthChange = (date) => {
+    if (requestAbortController.current) {
+      // make sure that you are aborting useless requests
+      // because it is possible to switch between months pretty quickly
+      requestAbortController.current.abort();
+    }
+
+    setIsLoading(true);
+    setHighlightedDays([]);
+    fetchHighlightedDays(date);
+  };
+
+  return (
+    <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <DateCalendar
+        defaultValue={initialValue}
+        loading={isLoading}
+        onMonthChange={handleMonthChange}
+        renderLoading={() => <DayCalendarSkeleton />}
+        slots={{
+          day: ServerDay,
+        }}
+        slotProps={{
+          day: {
+            highlightedDays,
+          },
+        }}
+      />
+    </LocalizationProvider>
   );
 }
